% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap.out.R
\name{capture.output.both}
\alias{capture.output.both}
\title{capture.output.both captures both output and messages}
\usage{
capture.output.both(
  ...,
  file = NULL,
  append = FALSE,
  type = c("both", "output", "message"),
  split = FALSE
)
}
\arguments{
\item{...}{Expressions to be evaluated}

\item{file}{A file name or a connection, or NULL to return the output as a character vector. If the connection is not open, it will be opened initially and closed on exit.}

\item{append}{logical. If file a file name or unopened connection, append or overwrite?}

\item{type}{is passed to sink, see there (but \code{both} is allowed)}

\item{split}{is passed to sink, see there}
}
\value{
A character string (if file = NULL), or invisible NULL
}
\description{
Adaptation of \code{\link[utils:capture.output]{utils::capture.output()}} to enable capture of both output and messages
}
\section{details}{


See \code{\link[utils:capture.output]{utils::capture.output()}} for detailed description. The current function allows \code{type='both'}.
}

