% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_child_chunk.R
\name{create_child_chunk}
\alias{create_child_chunk}
\title{create_child_chunk}
\usage{
create_child_chunk(contents, tmp_dir = ".")
}
\arguments{
\item{contents}{Character vector where each element is a line that will be included in the temporary file that will be created}

\item{tmp_dir}{Character string with the name of the folder that will contain the temporary file. Default: name of current folder}
}
\value{
Character string with the full name of the temporary file
}
\description{
\code{create_child_chunk} creates a temporary file that can be read into an \code{.Rmd} file as a child chunk.
}
\examples{
\dontrun{
# snippet of .Rmd file

```{r echo=F}
contents = readLines_part('codefile.R')
tfile = create_child(contents)
```

```{r child=tfile}
```

```{r echo=F}
unlink(tfile)
```
}
}
