% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_httr_get.R
\name{debug_httr_get}
\alias{debug_httr_get}
\title{debugging the httr::GET function}
\usage{
debug_httr_get(fn, ret = "finalurl")
}
\arguments{
\item{fn}{An expression that calls (indirectly) the \code{httr::GET} function}

\item{ret}{A character string or a numeric vector indicating the information that should be returned from the call(s) to \code{httr::GET}. Default: \code{finalurl}. See details section.}
}
\value{
List or character value with the requested information. See Details section
}
\description{
By temporarily setting a trace on \code{httr::GET} the \code{debug_httr_get} function is able to retrieve the url that is used as input and the resulting response object. This is especially useful when debugging a high-level function that somewhere in a subfunction makes one of more calls to \code{httr::GET}.
}
\section{Details}{


The function works by temporarily setting a trace on the \code{httr::GET} function. The result is a list with the requested information for each call to \code{httr::GET} that is caused by the expression in \code{fn}. If there is only one call to \code{httr::GET} the list has length 1 and will be unpacked. The possible values for \code{fn} and the returned information:
\describe{
\item{none or 'finalurl':}{the full url for which the \code{httr::GET} request is done}
\item{'inputurl':}{the url structure that was the input for \code{httr::GET} request}
\item{'resp':}{the resulting \code{httr::GET} response object}
\item{'json':}{the formatted \code{json} representation of the content of the response object}
\item{'fjson':}{the content of the response object after applying \code{jsonlite::fromJSON}}
\item{'org_result':}{the original results of the expression (as if no tracing had been used)}
\item{a numeric vector:}{a portion of the response object according to the vector used in \code{purrr::pluck} to extract from the response object}
}
}

\examples{
\dontrun{
x=debug_httr_get(  rtweet::get_trends("23424977")              )
print(x)
debug_httr_get(  rtweet::get_trends("23424977"), ret=c(9,2)  )
print(x)
x=debug_httr_get(  rtweet::get_trends("23424977"), ret='resp'  )
str(x)
}
}
