% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap.out.R
\name{display_wrapped}
\alias{display_wrapped}
\title{display_wrapped shows strings in the console}
\usage{
display_wrapped(strings, width = NULL, force_wrap = FALSE)
}
\arguments{
\item{strings}{A vector of strings}

\item{width}{The width (length) that each line will have (at most)}

\item{force_wrap}{Boolean When TRUE forces the use of \code{stringr::str_wrap}. When FALSE \code{HOQCutil::hard_split} will be used when \code{HOQCutil} is available (which is of course the case unless this function is copied outside the package). Default: FALSE}
}
\description{
Each string is split in pieces not exceeding a certain length by using \code{HOQCutil::hard_split} or \code{stringr::str_wrap}
}
\examples{
\dontrun{
x=glue::glue_collapse(rep(c(letters,' '),5))
display_wrapped(c(x,x),60)
display_wrapped(c(x,x),60,T)
}
}
