% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_WENS.R
\name{format_NS}
\alias{format_NS}
\title{format_NS}
\usage{
format_NS(x, N = "N", S = "S", dig = 2)
}
\arguments{
\item{x}{numeric vector with lattitude data}

\item{N}{character to use for positive lattitudes}

\item{S}{character to use for negative lattitudes}

\item{dig}{numeric scalar indicating number of digits after decimal point}
}
\description{
format_NS can be used to format lattitude data.
}
