% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_WENS.R
\name{format_WE}
\alias{format_WE}
\title{format_WE}
\usage{
format_WE(x, W = "W", E = "E", dig = 2)
}
\arguments{
\item{x}{numeric vector with longitude data}

\item{W}{character to use for negative longitudes}

\item{E}{character to use for positive longitudes}

\item{dig}{numeric scalar indicating number of digits after decimal point}
}
\description{
format_WE can be used to format longitude data.
}
