% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_odata4.R
\name{get_table_cbs_odata4_GET}
\alias{get_table_cbs_odata4_GET}
\title{get_table_cbs_odata4_GET : retrieves url}
\usage{
get_table_cbs_odata4_GET(url, ...)
}
\arguments{
\item{url}{Character string with url that will be passed to GET function without further encoding. It is assumed that a json result can be returned. Default: none}

\item{...}{Extra parameters. Only parameter currently allowed is response=TRUE. When this is used the httr response is returned and not a table. Useful for debugging.}
}
\value{
if not succesful a character string with an error message. If succesful the contents is regarded as a json object and translated to a data.frame or list when possible. If the extra parameter response=TRUE is set, the result is a httr response object.
}
\description{
This function handles the network IO for the \code{\link{get_table_cbs_odata4}} function but can also be used stand-alone.
}
\examples{
\dontrun{
myurl = "https://acc-ccb.cbs.nl/CBS/82931NED/Observations?$format=json&$skip=1&$top=2"
t1=get_table_cbs_odata4_GET(myurl)
t1=get_table_cbs_odata4_GET(myurl,response=T)
}
}
