% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap.out.R
\name{hard_split}
\alias{hard_split}
\title{hard_split splits the output of strings at a specific width}
\usage{
hard_split(strings, width = getOption("width", 110) - 3)
}
\arguments{
\item{strings}{A vector of strings}

\item{width}{The width (length) that each line will have (at most). Default \code{getOption('width',110)-3}}
}
\description{
Each string is split in pieces that have that length (or less for the last part of the string)
}
\section{acknowledgements}{

I was glad to be able to use the following (idea for) code from a StackOverflow article by \href{https://stackoverflow.com/questions/32398301/fastest-way-to-split-strings-into-fixed-length-elements-in-r}{akrun}.
}

\examples{

hard_split(paste(letters,collapse =''),width=10)

}
