% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igs.R
\name{igs}
\alias{igs}
\title{igs Insert graphical statement}
\usage{
igs(filename, opts = list(), level = 5)
}
\arguments{
\item{filename}{Character string with the name of the image file}

\item{opts}{List with options to add to \code{img} statement. See Details}

\item{level}{Integer indicating the level where the image file is located. E.g. \code{level=2} indicates that the location of the image file is \code{../../filename} relative to the document}
}
\value{
Character string with the generated html statement
}
\description{
generates html statement to insert graphics in a Hugo generated web document
}
\section{Details}{

Especially useful when an externally created image has to be included . It can be used in a text chunk with e.g.
`\verb{​r igs(...)}` . In \code{opts} can be specified all the options for the html \code{img} statement (with exception of \code{src}) :
\code{link}, \code{rel}, \code{alt}, \code{caption}, \code{width}, \code{height}, \code{attr} and \code{attrlink}
}

\examples{
\dontrun{
igs("post/2020-05-04-dir_files/my.png",opts=list(width='75\%',alt="my.png"))
# <p><img src="../../../../../post/2020-05-04-dir_files/my.png" width="75\%" alt="my.png" /></p>
}
}
