% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massign.R
\name{massign}
\alias{massign}
\alias{\%ma\%}
\title{assign values to multiple variables}
\usage{
massign(..., values = NULL, envir = parent.frame())

n \%ma\% v
}
\arguments{
\item{...}{One or more names that can be used as variable names.}

\item{values}{Values that are replicated where needed to assign to the variable names}

\item{envir}{The environment in which the values are assigned (bound) to the variable names}

\item{n}{Character vector with variable names.}

\item{v}{Values that are replicated where needed to assign to the variable names}
}
\description{
assign values to multiple variables
}
\section{acknowledgement}{

I encountered this function in a contribution by Tommy (\url{https://stackoverflow.com/users/662787/tommy}) in \url{https://stackoverflow.com/questions/7519790/assign-multiple-new-variables-on-lhs-in-a-single-line}. I adapted the function slightly to allow a character vector for the first argument and derived the binary operator \code{\%ma\%} from it. I also changed the name of the function from \code{vassign} to \code{massign} because the relevant feature is assignment to multiple variables.
}

\examples{
\dontrun{
massign(aa,bb, values= c(16,30) )
c('aa','bb') \%ma\% c(16,30)
}
}
