% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myknit.R
\name{myknit}
\alias{myknit}
\title{myknit documentation.}
\usage{
myknit(inputFile, encoding = getOption("encoding"), hoqc_render = TRUE)
}
\arguments{
\item{inputFile}{base file to render}

\item{encoding}{encoding of basefile}

\item{hoqc_render}{boolean to indicate that output file is to be created

This is the last line of the documentation.}
}
\description{
myknit documentation.
}
