% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PCA.R
\name{plot_PCA}
\alias{plot_PCA}
\title{plot_PCA: replacement for \link[FactoMineR:plot.PCA]{plot.PCA}}
\usage{
plot_PCA(res, var_ind = "ind", axes = 1:2, cex = 1, font.std = 1,
  pch.std = 20, col.std = "black", font.emp = 2, pch.emp = pch.std,
  col.emp = "red", font.sel = font.std, pch.sel = pch.std,
  col.sel = col.std, font.sup = 3, pch.sup = 1, col.sup = "blue",
  show.sup = TRUE, select = NULL, emphasize = NULL,
  invisible = NULL)
}
\arguments{
\item{res}{Structure that is the output of the function \link[FactoMineR:PCA]{PCA}}

\item{var_ind}{Character string that indicates if variables (\code{var}) or individuals (\code{ind}) will be plotted.  Default: \code{"ind"}}

\item{axes}{Numeric vector with two integers indicating the dimensions that will be plotted.  Default: \code{1:2}}

\item{cex}{Positive number indicating the size of symbols and labels. Default: \code{1}}

\item{font.std}{Positive integer indicating the font that is used for the standard labels (i.e. not for supplemental, selected or emphasized elements). See description of \code{par()} for details: 1 plain text, 2 bold face, 3 italic and 4 bold italic. Default: \code{1}}

\item{pch.std}{Positive integer indicating the plotting character that is used for the standard points (i.e. not for supplemental, selected or emphasized elements). See description of \code{points} for details: 0 open square, 1 open circle, 2 open triangle etc. Default: \code{20} (i.e. very small closed circle)}

\item{col.std}{Color specification for the standard elements (i.e. not for supplemental, selected or emphasized elements). See description of \code{par()} for details of specification. Default: \code{"black"}}

\item{font.emp}{Positive integer indicating the font that is used for the labels of emphasized elements. See description of \code{font.std}. Default: \code{2}}

\item{pch.emp}{Positive integer indicating the plotting character that is used for the emphasized points. See description of \code{pch.std}. Default: \code{pch.std}}

\item{col.emp}{Color specification for emphasized elements. See description of \code{col.std}. Default: \code{"red"}}

\item{font.sel}{Positive integer indicating the font that is used for the labels of selected elements. See description of \code{font.std}. Default: \code{font.std}}

\item{pch.sel}{Positive integer indicating the plotting character that is used for the selected points. See description of \code{pch.std}. Default: \code{pch.std}}

\item{col.sel}{Color specification for selected elements. See description of \code{col.std}. Default: \code{col.std}}

\item{font.sup}{Positive integer indicating the font that is used for the labels of supplemental elements. See description of \code{font.std}. Default: \code{3}}

\item{pch.sup}{Positive integer indicating the plotting character that is used for the supplemental points. See description of \code{pch.std}. Default: \code{1}}

\item{col.sup}{Color specification for supplemental elements. See description of \code{col.std}. Default: \code{"blue"}}

\item{show.sup}{Boolean indicating if supplemental elements should be shown. Default: \code{TRUE}}

\item{select}{Character string  with a selection criterion (see \code{details}) or an integer vector with the sequence numbers of the active elements that are selected (i.e. be given a different layout than the other elements). Default: NULL (i.e. all are selected)}

\item{emphasize}{Character vector with the rownames or an integer vector with the sequence numbers of the active elements that should be emphasized (i.e. be given a different layout than the other elements). Default: NULL (i.e. none are emphasized)}

\item{invisible}{Character vector with the rownames of the elements (active and supplemental) that are not shown. Default: NULL}
}
\description{
plot_PCA is a replacement for \link[FactoMineR:plot.PCA]{plot.PCA}. Out of this function I took the functionality that I need and added the functionality to emphasize points (variables or individuals).
The function plots the variables or individuals information from the structure returned by the function \link[FactoMineR:PCA]{PCA}.
}
\section{Details}{

The select mechanism is taken over asis from the \code{FactoMineR::PCA} function. The labels of selected elements will be plotted.\cr The authors Francois Husson \url{husson@agrocampus-ouest.fr} and Jeremy Mazet describe it as follows:\cr\cr
The select argument can be used in order to select a part of the elements (individuals if you draw the graph of individuals, or variables if you draw the graph of variables) that are drawn. For example, you can use:\cr
\code{select = 1:5} and then the elements \code{1:5} are drawn.\cr
\code{select = c("name1","name5")} and then the elements that have the names \code{name1} and \code{name5} are drawn.\cr
\code{select = "coord 10"} and then the 10 elements that have the highest (squared) coordinates on the 2 chosen dimensions are drawn.\cr
\code{select = "contrib 10"} and then the 10 elements that have the highest contribution on the 2 dimensions of your plot are drawn.\cr
\code{select = "cos2 5"} and then the 5 elements that have the highest cos2 on the 2 dimensions of your plot are drawn.\cr
\code{select = "dist 8"} and then the 8 elements that have the highest distance to the center of gravity are drawn.\cr\cr
NB. the following functionality is contained in \link[FactoMineR]{plot.PCA} but not described there:\cr
\code{select = "cos2 0.8"} and then the elements for which the sum of the cos2 on the 2 dimensions of your plot is greater than 0.8 are drawn.
}

\section{Acknowledgements}{

All ideas come from the creators of the \link[FactoMineR:FactoMineR-package]{FactoMineR} package. This implementation is mine and if you find any errors, these will be mine. In case of errors try using the original \link[FactoMineR:plot.PCA]{plot.PCA}.
}

\examples{
\dontrun{
plot_PCA(res,cex=0.8,show.sup = F,emphasize = 'BU03620304',sel='contrib 6')
}
}
