% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PCA.R
\name{print_PCA}
\alias{print_PCA}
\title{print_PCA: prints a part of \link[FactoMineR]{PCA} output (sorted)}
\usage{
print_PCA(res, var_ind = "ind", measure = "contrib", decreasing = T,
  absolute = T, dim = 1)
}
\arguments{
\item{res}{Structure that is the output of the function \link[FactoMineR]{PCA}}

\item{var_ind}{Character string that indicates if output for variables (\code{var}) or individuals (\code{ind}) will be printed.  Default: \code{'ind'}}

\item{measure}{Character string with the name of the matrix. Possible values are  \code{'coord'}, \code{'cor'}, \code{'cos2'}, \code{'contrib'} and \code{'dist'}. Default: \code{'contrib'}}

\item{decreasing}{Boolean indicating that greatest values come first (TRUE) or smallest values come first (FALSE). Default: \code{TRUE}}

\item{absolute}{Boolean indicating that that absolute values will be sorted: TRUE means take absolute value and FALSE means leave value asis. (Only relevant for\code{measuer='coord'}). Default: \code{TRUE}}

\item{dim}{Integer (or vector of two integers) indicating the dimension(s) on which the matrix will be sorted. Default: \code{1}}
}
\description{
The output of \code{\link[FactoMineR]{PCA}} of the \link\link{FactoMineR} package is an object that contains various matrices. The function \code{\link[FactoMineR]{print.PCA}} of that  package gives an overview of the available matrices. This function displays a sorted version of one of these matrices. So it is easier to check which variables or individuals play a large role for a certain dimension or pair of dimensions.
}
\examples{
\dontrun{
print_PCA(res,var_ind = 'var',decreasing = T,
      dim=1,measure = 'coord',abs=T)
}
}
