% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxtable.R
\name{pxtable}
\alias{pxtable}
\title{print data.frame in LaTeX}
\usage{
pxtable(df, tablabel = "", tabcap = "", adef = "l", ap = c(),
  av = c(), ddef = 0, dp = c(), dv = c(), tco = T, tcf = tco,
  te = F, newpage = F, scalebox = NULL, include.colnames = T,
  colnames = NA, rotate.colnames = T, include.rownames = F,
  sanitize.colnames.function = function(x) x, booktabs = F, ...)
}
\arguments{
\item{df}{data.frame that will be printed.}

\item{tablabel}{Character string that can be used to reference the table as \code{table:tablabel}. Default: ''}

\item{tabcap}{Character string that is as caption for the table. Default: ''}

\item{adef}{Character string with the default alignment. Default: "l"}

\item{ap}{Integer vector with the positions of the columns for which an alignment specification is given in \code{av}. Default: c()}

\item{av}{Character vector with aligment specifications for the positions in \code{ap}. E.g. with values as 'c', 'r', 'p{3cm}' . Default: c()}

\item{ddef}{Integer with the default number of digits to display. Default: 0}

\item{dp}{Integer vector with the positions of the columns for which an digit specification is given in \code{dv}. Default: c()}

\item{dv}{Integer vector with digit specifications for the positions in \code{dp}. Default: c()}

\item{tco}{Boolean indicating if 'continued on' should be displayed at the bottom of a page(only) when using the \code{longtable} package. Default: T}

\item{tcf}{Boolean indicating if 'continued from' should be displayed at the top of a page(only) when using the \code{longtable} package. Default: \code{tco}}

\item{te}{Boolean indicating if 'end of table' should be displayed at the end of the table (only) when using the \code{longtable} package. Default: F}

\item{newpage}{Boolean indicating if the table should start on a new page. Default: F}

\item{scalebox}{Positive number for scaling the table. Forces the \code{tabular} environment. See \code{xtable::print.xtable}. Default: NULL}

\item{include.colnames}{Boolean indicating if column names will be printed. See \code{xtable::print.xtable}. Default: T}

\item{colnames}{Character vector with column headers to use instead of field names. Default: NA}

\item{rotate.colnames}{Boolean indicating if column names will be printed vertically. See \code{xtable::print.xtable}. Default: T}

\item{include.rownames}{Boolean indicating if row names will be printed. See \code{xtable::print.xtable}. Default: F}

\item{sanitize.colnames.function}{Function used for nicer printing of column names. See \code{xtable::print.xtable}. Default: no change in column names}

\item{booktabs}{Boolean indicating if LaTeX package \code{booktabs} will be used for formatting horizontal lines. See \code{xtable::print.xtable}. Default: F}

\item{...}{Additional arguments that are passed to \code{xtable::print.xtable}.}
}
\value{
See \code{xtable::print.xtable} as this function prepares only its function arguments. Using this function in a \code{knitr} chunk with \code{results ='markup'} will show the generated LaTeX text. With \code{results ='asis'} the text is transfered to the intermediate output file and interpreted as LaTeX in the final pdf document.
}
\description{
This function adds to the print function of the \code{xtable} package by providing predefined \code{add.to.row} code for the \code{longtable} LaTeX package: 'continued on/from ...' and 'end of table' lines that can switched on or off. This functionality can only be used with the \code{longtable} environment. When \code{scalebox} is specified the \code{tabular} environment is used and \code{add.to.row} code is not available. If the \code{longtable} package is not available then set \code{scalebox=1} or another appropriate value
}
\examples{
\dontrun{
df1 = data.frame(
 v1 = rep(c(1,NA),50),
 v2 = rep(c('R version 3.5.0', 'R is free software'),50),
 v3 = rep(c('Copyright (C) 2018', 'You are welcome'),50),
 v4 = rep(c('a','b'),50),
 v5 = rep(c(pi,NA),50),
 stringsAsFactors = F
)

HOQCutil::pxtable  (df1,tablabel='tab1',tabcap='mycaption',
	tablabel = "tab1", tabcap = "mycaption",
	ap = 1:4, av = c("r", "p{6cm}", "p{6cm}", "c"), adef = "l",
	dp = c(1, 5), dv = c(2, 3), ddef = 0, te = F,
	include.rownames = T, rotate.rownames = T
}
}
