% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pdf_funs.R
\name{read_pdf}
\alias{read_pdf}
\title{read text of pdf-file by line or by cell}
\usage{
read_pdf(filename, vtolerance = 2, by = "line", pageno = 1)
}
\arguments{
\item{filename}{Character string with path of the pdf-file}

\item{vtolerance}{Numeric scalar with vertical tolerance}

\item{by}{Character string with value "line" or "cell" indicating if text is gathered by text line of text cell}

\item{pageno}{Integer indicating the number of the page to read}
}
\value{
data.frame with the fields 'text', 'y'  and (when by=="cell")  'x'
}
\description{
read text of pdf-file by line or by cell
}
\section{Details}{

Uses \code{\link[pdftools:pdf_data]{pdftools::pdf_data}} as workhorse
}

\examples{
\dontrun{
df1 <- read_pdf (r"(D:\data\R\TTVA\inputs\TTV Amstelveen teamindeling Senioren VJ22.pdf)", by= "line")
names(df1) # [1] "y" "text"
df1 <- read_pdf (r"(D:\data\R\TTVA\inputs\TTV Amstelveen teamindeling Senioren VJ22.pdf)", by= "cell")
names(df1) # [1] "x"    "text" "y"
}

}
