% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pdf_funs.R
\name{read_pdf_fields}
\alias{read_pdf_fields}
\title{read text of a table from a page of a pdf-file}
\usage{
read_pdf_fields(
  filename,
  vtolerance = 2,
  htolerance = 2,
  header_line = 1,
  pageno = 1
)
}
\arguments{
\item{filename}{Character string with path of the pdf-file}

\item{vtolerance}{Numeric scalar with vertical tolerance to fix vertical mismatches}

\item{htolerance}{Numeric scalar with horizontal tolerance to fix mismatches field contents (field starts before header)}

\item{header_line}{Integer indicating which lines contain the headers of the table}

\item{pageno}{Integer indicating the number of the page to read}
}
\value{
data.frame with the table . All fields have character values
}
\description{
read text of a table from a page of a pdf-file
}
\section{Details}{

It is assumed that the table occupies a whole page and that the columns are defined by the words in the header.\cr
In the following example
\preformatted{field1         field2           field3}
\preformatted{v1a v1b        v2a  v2b     v2c  v3a   v3b}
field1 will be filled with "v1a v1b", field2 with "v2a v2b" and field3 with "v3a v3b".\cr
Multiple words in a field are separated by only one blank (even when the original data contains more than one blank)\cr\cr
The actual reading of the file is done with  \code{\link{read_pdf}} and  \code{\link[pdftools:pdf_data]{pdftools::pdf_data}} .
}

\examples{
\dontrun{
df1 <- read_pdf_fields (r"(D:\data\R\TTVA\inputs\TTV Amstelveen teamindeling Senioren VJ22.pdf)" )
names(df1) # [1] "Teamnr."    "Klasse"     "Teamrating" "Captain"    "Speler"     "Rating"     "Speeldag"
}

}
