% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_fns.R
\name{scan_with_hocr}
\alias{scan_with_hocr}
\title{doing a Tesseract scan with HOCR output}
\usage{
scan_with_hocr(
  img,
  confsel = F,
  extract_bbox = T,
  add_header_cols = F,
  engine = tesseract::tesseract("eng")
)
}
\arguments{
\item{img}{An image object or a character string with the name of an image file}

\item{confsel}{A Boolean indicating if the confidence rate should also be selected}

\item{extract_bbox}{A Boolean indicating if the bounding box should be unpacked (into x and y coordinates)}

\item{add_header_cols}{A Boolean indicating if two header columns (\code{header_col} and \code{header_col_seq}) should be added to the result and initialized to resp. \code{0} and \code{1}. Useful when \code{\link[=extract_table]{extract_table()}} is used later on.}

\item{engine}{The OCR engine to use. See \code{\link[tesseract:tesseract]{tesseract::tesseract()}}}
}
\value{
A data.frame with the scanned words. See \strong{Details}
}
\description{
A Tesseract scan with HOCR output returns an XHTML document with not only the scanned word, but also information about the line where the word is found and the bounding box. The function \code{scan_with_hocr} does the scan and converts the document to a data frame. See \strong{Details} and \strong{Acknowledgment}.
}
\section{Details}{

The result is a data.frame with one row for each word found and the following columns
\itemize{
\item line : the line on which the word was found
\item fldnr: the sequence number of the word on this line
\item word : the word that is recognized by the engine
\item bbox : the bounding box where the word was found (character string with e.g. '19 227 1087 251' indicating x-coordinates x1=19 and x2=1087 and y-coordinates y1=227 and y2=251). Not present when \code{extract_bbox=T} is set: in that case x1, x2, y1 and y2 are present.
\item conf : the confidence rate of the word (only when \code{confsel=T} is set)
\item header_col : column with \code{0}-s (only when \code{add_header_cols=T} is set)
\item header_col_seq : column with \code{1}-s (only when \code{add_header_cols=T} is set)
}
}

\section{Acknowledgment}{

This function is an extension of the snippet by \href{https://ropensci.org/technotes/2018/02/14/tesseract-18/}{Jeroen Ooms}. I only added the extraction of the line info. Afterwards I made it into a function and usable for connecting it with \code{\link[=extract_table]{extract_table()}}.
}

\examples{
\dontrun{
df1  = scan_with_hocr(img2,add_header_cols=F)
}
}
\seealso{
\link{scanner_functions} , \code{\link[=cleanup_bw]{cleanup_bw()}} and \code{\link[=extract_table]{extract_table()}}
}
