% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_fns.R
\name{scanner_functions}
\alias{scanner_functions}
\title{Scanner functions}
\description{
\code{\link[=cleanup_bw]{cleanup_bw()}}, \code{\link[=scan_with_hocr]{scan_with_hocr()}} and \code{\link[=extract_table]{extract_table()}} can be used to cleanup and scan (OCR) an image and extract a table into a data.frame format. The workflow would then be:
\itemize{
\item read an image from file with \code{\link[magick:editing]{magick::image_read()}} \cr e.g.
\code{img1 = magick::image_read('example1.png')}
\item define the list with cleanup options \cr e.g.
\code{cln_options1 = list(resize="4000x",trim=10,enhance=TRUE,sharpen=1)}
\item use the \code{\link[=cleanup_bw]{cleanup_bw()}} function with this list \cr e.g.
\code{img2 = cleanup_bw (img1,cln_options1)}
\item scan (OCR) the cleansed image with \code{\link[=scan_with_hocr]{scan_with_hocr()}} \cr e.g.
\code{df1  = scan_with_hocr(img2,add_header_cols=F)}
\item indicate in the columns of \code{df1} which fields belong to the table headers (or alternatively define a \code{headers} list)
\item extract the table with the \code{\link[=extract_table]{extract_table()}} function \cr e.g.
\code{df2= extract_table(df1, headers=NULL,lastline = Inf, desc_above=T)} or alternatively \cr
\code{df2= extract_table(df1, headers=headers,lastline = Inf, desc_above=T)}
}
}
