% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pdf_funs.R
\name{text2pdf}
\alias{text2pdf}
\title{write text to pdf file}
\usage{
text2pdf(txt, pdffile)
}
\arguments{
\item{txt}{character vector with data to write to pdffile}

\item{pdffile}{character string with name (optionally with full path) of pdf file}
}
\description{
write text to pdf file
}
\section{Details}{

\code{text2pdf} writes \code{txt} to a temporary text file with an additional line
added before and after \code{txt} with three backticks.
\code{pandoc} then considers the file as verbatim text and converts it to a pdf file
}

\examples{
\dontrun{
outfile <- "testfile.pdf"
txt     <- c('en dat is één', 'en dat is twee', 'en dat is zev-e-ven')
v <- text2pdf (txt, outfile)
}

}
