% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_odata4.R
\name{get_table_cbs_odata4}
\alias{get_table_cbs_odata4}
\title{get_table_cbs_odata4 : retrieves CBS Odata4 table data}
\usage{
get_table_cbs_odata4(root = NULL, table_id = NULL, subtable = NULL,
  query = NULL, verbose = F, encode = T,
  odata_root = "https://acc-ccb.cbs.nl", odata_cat = "CBS", ...)
}
\arguments{
\item{table_id}{Character string identifying table for which information will be returned. When table_id == NULL catalog information will be returned. Default: NULL}

\item{subtable}{Character string indicating subtable for which information will be returned. When subtable == NULL information about the available subtables will be returned. Default: NULL}

\item{query}{Character string with an OData4 query. See \url{http://docs.oasis-open.org/odata/odata/v4.0/odata-v4.0-part1-protocol.html} for the general OData4 query possibilities and \url{https://acc-ccb.cbs.nl/implement.html} for the subset of the CBS implementation. Default: ""}

\item{verbose}{Boolean indicating if the generated url should be shown as a message. Default: FALSE}

\item{encode}{Boolean indicating if the query has to be encode by URLencode. Default: TRUE}

\item{odata_root}{Character string with the root for the url. Default: "https://acc-ccb.cbs.nl"}

\item{odata_cat}{Character string with the catalog identifier. Default: "CBS"}

\item{...}{Parameters to be passed to get_table_cbs_odata4_GET. Only parameter currently allowed is response=TRUE. When this is used the httr response is returned and not a table. Useful for debugging.}
}
\value{
if not succesful a character string with an error message. If succesful a data.frame when the contents of a subtable was requested and a list with the properties of the table when Properties was requested. If the extra parameter response=TRUE is set, the result is a httr response object.
}
\description{
This function can be used to retrieve data or information from the new \href{https://www.cbs.nl}{CBS} ('Centraal Bureau voor de Statistiek' or 'Statistics Netherlands') Odata4 data infrastructure. The general OData4 protocol is described in \url{http://docs.oasis-open.org/odata/odata/v4.0/odata-v4.0-part1-protocol.html} and the part that is implemented by CBS in  \url{https://acc-ccb.cbs.nl/implement.html}
}
\examples{
\dontrun{
t1=get_table_cbs_odata4()
t1=get_table_cbs_odata4(subtable="Datasets")
t1=get_table_cbs_odata4(subtable="Catalogs")
t1=get_table_cbs_odata4(table_id="82931NED")
t1=get_table_cbs_odata4(table_id="82931NED",subtable="Dimensions")
t1=get_table_cbs_odata4(table_id="82931NED",subtable="Properties")
t1=get_table_cbs_odata4(table_id="82931NED",subtable="Observations",query="$skip=1&$top=2")
t1=get_table_cbs_odata4(table_id="82931NED",subtable="Observations",query="$skip=1&$top=2&$select=Measure,Value,WijkenEnBuurten") # does not work
}
}
