% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silent_library.R
\name{silent_library}
\alias{silent_library}
\title{attach library as silent as possible}
\usage{
silent_library(package_name, mywarnings = FALSE)
}
\arguments{
\item{package_name}{Character string with name of package or \code{name} (as in \code{as.name}) of package}

\item{mywarnings}{Boolean indicating if warnings should be displayed. Default: FALSE}
}
\value{
Character vector (invisible) with the names of attached packages after attaching \code{package_name}: identical to (invisible) .packages()
}
\description{
attach library as silent as possible
}
\examples{
\dontrun{
silent_library('knitr')
silent_library(knitr)
packages_to_load <- c('glue','FactoMineR','curl')
purrr::walk(packages_to_load[1:2],silent_library)
silent_library(packages_to_load[[3]])
}
}
